+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Remove AM/PM from imessage timestamp strings
 
 # 1.1.0
 * Chronological emails page